+++ tencentcloud/gaap/v20180529/models.go
@@ -3341,6 +3341,57 @@ func (r *DescribeAccessRegionsResponse) FromJsonString(s string) error {
 }
 
 // Predefined struct for user
+type DescribeAuthSignatureRequestParams struct {
+
+}
+
+type DescribeAuthSignatureRequest struct {
+	*tchttp.BaseRequest
+	
+}
+
+func (r *DescribeAuthSignatureRequest) ToJsonString() string {
+    b, _ := json.Marshal(r)
+    return string(b)
+}
+
+// FromJsonString It is highly **NOT** recommended to use this function
+// because it has no param check, nor strict type check
+func (r *DescribeAuthSignatureRequest) FromJsonString(s string) error {
+	f := make(map[string]interface{})
+	if err := json.Unmarshal([]byte(s), &f); err != nil {
+		return err
+	}
+	
+	if len(f) > 0 {
+		return tcerr.NewTencentCloudSDKError("ClientError.BuildRequestError", "DescribeAuthSignatureRequest has unknown keys!", "")
+	}
+	return json.Unmarshal([]byte(s), &r)
+}
+
+// Predefined struct for user
+type DescribeAuthSignatureResponseParams struct {
+	// 唯一请求 ID，每次请求都会返回。定位问题时需要提供该次请求的 RequestId。
+	RequestId *string `json:"RequestId,omitempty" name:"RequestId"`
+}
+
+type DescribeAuthSignatureResponse struct {
+	*tchttp.BaseResponse
+	Response *DescribeAuthSignatureResponseParams `json:"Response"`
+}
+
+func (r *DescribeAuthSignatureResponse) ToJsonString() string {
+    b, _ := json.Marshal(r)
+    return string(b)
+}
+
+// FromJsonString It is highly **NOT** recommended to use this function
+// because it has no param check, nor strict type check
+func (r *DescribeAuthSignatureResponse) FromJsonString(s string) error {
+	return json.Unmarshal([]byte(s), &r)
+}
+
+// Predefined struct for user
 type DescribeBlackHeaderRequestParams struct {
 
 }