+++ tencentcloud/lighthouse/v20200324/errors.go
@@ -20,6 +20,9 @@ const (
 	// CAM签名/鉴权错误。
 	AUTHFAILURE = "AuthFailure"
 
+	// 该地域尚未开放，请选择其他地域。
+	AUTHFAILURE_INVALIDREGION = "AuthFailure.InvalidRegion"
+
 	// 操作失败。
 	FAILEDOPERATION = "FailedOperation"
 
@@ -50,6 +53,9 @@ const (
 	// 查询流量包失败。
 	FAILEDOPERATION_DESCRIBEINSTANCESTRAFFICPACKAGESFAILED = "FailedOperation.DescribeInstancesTrafficPackagesFailed"
 
+	// 查询资源返回了不符合要求内容。
+	FAILEDOPERATION_DESCRIBERESOURCESRETURNABLEERROR = "FailedOperation.DescribeResourcesReturnableError"
+
 	// 销毁资源失败，请稍后重新操作。
 	FAILEDOPERATION_DESTROYRESOURCESFAILED = "FailedOperation.DestroyResourcesFailed"
 
@@ -61,6 +67,9 @@ const (
 
 	// 对实例的操作失败。
 	FAILEDOPERATION_INSTANCEOPERATIONFAILED = "FailedOperation.InstanceOperationFailed"
+
+	// 命令无法找到。
+	FAILEDOPERATION_INVALIDCOMMANDNOTFOUND = "FailedOperation.InvalidCommandNotFound"
 
 	// 退还资源失败。
 	FAILEDOPERATION_ISOLATERESOURCESFAILED = "FailedOperation.IsolateResourcesFailed"
@@ -79,6 +88,9 @@ const (
 
 	// 快照操作失败。
 	FAILEDOPERATION_SNAPSHOTOPERATIONFAILED = "FailedOperation.SnapshotOperationFailed"
+
+	// 调用计费网关服务失败，请稍后重新操作操作。
+	FAILEDOPERATION_TRADECALLBILLINGGATEWAYFAILED = "FailedOperation.TradeCallBillingGatewayFailed"
 
 	// 计费询价失败。
 	FAILEDOPERATION_TRADEGETPRICEFAILED = "FailedOperation.TradeGetPriceFailed"
@@ -143,6 +155,9 @@ const (
 	// 产品未定义的套餐 ID。
 	INVALIDPARAMETER_BUNDLEIDNOTFOUND = "InvalidParameter.BundleIdNotFound"
 
+	// 参数冲突。
+	INVALIDPARAMETER_CONFLICT = "InvalidParameter.Conflict"
+
 	// 参数非法，Filter 参数中的 Values 取值数量超过允许的最大数量。
 	INVALIDPARAMETER_FILTERVALUELIMITEXCEEDED = "InvalidParameter.FilterValueLimitExceeded"
 
@@ -170,6 +185,9 @@ const (
 	// 参数非法，Filter 参数中有不支持的 Name。
 	INVALIDPARAMETER_INVALIDFILTERNOTSUPPORTEDNAME = "InvalidParameter.InvalidFilterNotSupportedName"
 
+	// 必须要指定一个要修改的属性。
+	INVALIDPARAMETER_MUSTSPECIFYONEATTRIBUTETOMODIFY = "InvalidParameter.MustSpecifyOneAttributeToModify"
+
 	// 参数非法，每次只能修改一个属性。
 	INVALIDPARAMETER_ONLYALLOWMODIFYONEATTRIBUTE = "InvalidParameter.OnlyAllowModifyOneAttribute"
 
@@ -194,8 +212,14 @@ const (
 	// 云联网实例ID格式非法。
 	INVALIDPARAMETERVALUE_CCNIDMALFORMED = "InvalidParameterValue.CcnIdMalformed"
 
+	// 客户令牌长度超出限制。
+	INVALIDPARAMETERVALUE_CLIENTTOKENTOOLONG = "InvalidParameterValue.ClientTokenTooLong"
+
 	// 参数值非法，云硬盘备份点 ID 格式非法。
 	INVALIDPARAMETERVALUE_DISKBACKUPIDMALFORMED = "InvalidParameterValue.DiskBackupIdMalformed"
+
+	// 指定的云盘备份点名称不可大于最大长度。
+	INVALIDPARAMETERVALUE_DISKBACKUPNAMETOOLONG = "InvalidParameterValue.DiskBackupNameTooLong"
 
 	// 磁盘名称长度超出限制。
 	INVALIDPARAMETERVALUE_DISKNAMETOOLONG = "InvalidParameterValue.DiskNameTooLong"
@@ -320,6 +344,9 @@ const (
 	// 实例挂载数据盘配额不足，无法挂载磁盘。
 	LIMITEXCEEDED_ATTACHDATADISKQUOTALIMITEXCEEDED = "LimitExceeded.AttachDataDiskQuotaLimitExceeded"
 
+	// 配额不足，当前自定义镜像配额不允许创建新的自定义镜像。
+	LIMITEXCEEDED_BLUEPRINTQUOTALIMITEXCEEDED = "LimitExceeded.BlueprintQuotaLimitExceeded"
+
 	// 超过磁盘备份点配额限制。
 	LIMITEXCEEDED_DISKBACKUPQUOTALIMITEXCEEDED = "LimitExceeded.DiskBackupQuotaLimitExceeded"
 
@@ -367,6 +394,9 @@ const (
 
 	// 禁止对实例进行操作，实例最近一次的操作尚在进行中。
 	OPERATIONDENIED_INSTANCEOPERATIONINPROGRESS = "OperationDenied.InstanceOperationInProgress"
+
+	// 禁止创建快照。
+	OPERATIONDENIED_OPERATIONDENIEDCREATESNAPSHOT = "OperationDenied.OperationDeniedCreateSnapshot"
 
 	// 使用存储型套餐的实例不支持创建快照。
 	OPERATIONDENIED_OPERATIONDENIEDCREATESNAPSHOTFORSTORAGEBUNDLE = "OperationDenied.OperationDeniedCreateSnapshotForStorageBundle"