+++ tencentcloud/tdmq/v20200217/models.go
@@ -2225,10 +2225,10 @@ type CreateRocketMQNamespaceRequestParams struct {
 	// 命名空间名称，3-64个字符，只能包含字母、数字、“-”及“_”
 	NamespaceId *string `json:"NamespaceId,omitempty" name:"NamespaceId"`
 
-	// 未消费消息的保留时间，以毫秒为单位，60秒-15天
+	// 已废弃
 	Ttl *uint64 `json:"Ttl,omitempty" name:"Ttl"`
 
-	// 消息持久化后保留的时间，以毫秒为单位
+	// 已废弃
 	RetentionTime *uint64 `json:"RetentionTime,omitempty" name:"RetentionTime"`
 
 	// 说明，最大128个字符
@@ -2244,10 +2244,10 @@ type CreateRocketMQNamespaceRequest struct {
 	// 命名空间名称，3-64个字符，只能包含字母、数字、“-”及“_”
 	NamespaceId *string `json:"NamespaceId,omitempty" name:"NamespaceId"`
 
-	// 未消费消息的保留时间，以毫秒为单位，60秒-15天
+	// 已废弃
 	Ttl *uint64 `json:"Ttl,omitempty" name:"Ttl"`
 
-	// 消息持久化后保留的时间，以毫秒为单位
+	// 已废弃
 	RetentionTime *uint64 `json:"RetentionTime,omitempty" name:"RetentionTime"`
 
 	// 说明，最大128个字符
@@ -8681,10 +8681,10 @@ type ModifyRocketMQNamespaceRequestParams struct {
 	// 命名空间名称，3-64个字符，只能包含字母、数字、“-”及“_”
 	NamespaceId *string `json:"NamespaceId,omitempty" name:"NamespaceId"`
 
-	// 未消费消息的保留时间，以毫秒为单位，60秒-15天
+	// 已废弃
 	Ttl *uint64 `json:"Ttl,omitempty" name:"Ttl"`
 
-	// 消息持久化后保留的时间，以毫秒为单位
+	// 已废弃
 	RetentionTime *uint64 `json:"RetentionTime,omitempty" name:"RetentionTime"`
 
 	// 说明，最大128个字符
@@ -8700,10 +8700,10 @@ type ModifyRocketMQNamespaceRequest struct {
 	// 命名空间名称，3-64个字符，只能包含字母、数字、“-”及“_”
 	NamespaceId *string `json:"NamespaceId,omitempty" name:"NamespaceId"`
 
-	// 未消费消息的保留时间，以毫秒为单位，60秒-15天
+	// 已废弃
 	Ttl *uint64 `json:"Ttl,omitempty" name:"Ttl"`
 
-	// 消息持久化后保留的时间，以毫秒为单位
+	// 已废弃
 	RetentionTime *uint64 `json:"RetentionTime,omitempty" name:"RetentionTime"`
 
 	// 说明，最大128个字符