+++ tencentcloud/trtc/v20190722/models.go
@@ -130,7 +130,7 @@ type CreateCloudRecordingRequestParams struct {
 	// 云端录制控制参数。
 	RecordParams *RecordParams `json:"RecordParams,omitempty" name:"RecordParams"`
 
-	// 云端录制文件上传到云存储的参数(目前支持云点播VOD和对象存储COS)。
+	// 云端录制文件上传到云存储的参数(目前支持云点播VOD和对象存储COS)。点播和对象存储的参数必填其中之一，不支持同时设置点播和对象存储。
 	StorageParams *StorageParams `json:"StorageParams,omitempty" name:"StorageParams"`
 
 	// TRTC房间号的类型，必须和录制的房间所对应的RoomId类型相同:
@@ -169,7 +169,7 @@ type CreateCloudRecordingRequest struct {
 	// 云端录制控制参数。
 	RecordParams *RecordParams `json:"RecordParams,omitempty" name:"RecordParams"`
 
-	// 云端录制文件上传到云存储的参数(目前支持云点播VOD和对象存储COS)。
+	// 云端录制文件上传到云存储的参数(目前支持云点播VOD和对象存储COS)。点播和对象存储的参数必填其中之一，不支持同时设置点播和对象存储。
 	StorageParams *StorageParams `json:"StorageParams,omitempty" name:"StorageParams"`
 
 	// TRTC房间号的类型，必须和录制的房间所对应的RoomId类型相同: